Update HW1_1.17.c
I tried use the exponential form of my number

+++ HW1_1.17.c
@@ -2,12 +2,12 @@
 #include <math.h>
 
 int main() {
-   double x, x_8;
+    long double x, x_8;
    printf("enter a number within 1 to 10000:");
-   scanf("%lf", &x);
+   scanf("%Lf", &x);
    if (x <= 10000 && x >= 0) {
        x_8 = pow(x, 8);
-       printf("%20.4lf\n", x_8);
+       printf("%20.4Le\n", x_8);
    }
    else {
         printf("invalid input, enter a number within 1 to 10000");
